fix namespace assignment not passing through
