+++ cmd/helm/helm.go
@@ -17,35 +17,19 @@ limitations under the License.
 package main // import "helm.sh/helm/v4/cmd/helm"
 
 import (
-	"io"
 	"log"
 	"os"
-	"strings"
-
-	"github.com/spf13/cobra"
-	"sigs.k8s.io/yaml"
 
 	// Import to initialize client auth plugins.
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 
 	"helm.sh/helm/v4/pkg/action"
-	"helm.sh/helm/v4/pkg/cli"
 	helmcmd "helm.sh/helm/v4/pkg/cmd"
 	"helm.sh/helm/v4/pkg/kube"
-	kubefake "helm.sh/helm/v4/pkg/kube/fake"
-	release "helm.sh/helm/v4/pkg/release/v1"
-	"helm.sh/helm/v4/pkg/storage/driver"
 )
-
-var settings = cli.New()
 
 func init() {
 	log.SetFlags(log.Lshortfile)
-}
-
-// hookOutputWriter provides the writer for writing hook logs.
-func hookOutputWriter(_, _, _ string) io.Writer {
-	return log.Writer()
 }
 
 func main() {
@@ -62,17 +46,6 @@ func main() {
 		os.Exit(1)
 	}
 
-	cobra.OnInitialize(func() {
-		helmDriver := os.Getenv("HELM_DRIVER")
-		if err := actionConfig.Init(settings.RESTClientGetter(), settings.Namespace(), helmDriver, helmcmd.Debug); err != nil {
-			log.Fatal(err)
-		}
-		if helmDriver == "memory" {
-			loadReleasesInMemory(actionConfig)
-		}
-		actionConfig.SetHookOutputFunc(hookOutputWriter)
-	})
-
 	if err := cmd.Execute(); err != nil {
 		helmcmd.Debug("%+v", err)
 		switch e := err.(type) {
@@ -82,42 +55,4 @@ func main() {
 			os.Exit(1)
 		}
 	}
-}
-
-// This function loads releases into the memory storage if the
-// environment variable is properly set.
-func loadReleasesInMemory(actionConfig *action.Configuration) {
-	filePaths := strings.Split(os.Getenv("HELM_MEMORY_DRIVER_DATA"), ":")
-	if len(filePaths) == 0 {
-		return
-	}
-
-	store := actionConfig.Releases
-	mem, ok := store.Driver.(*driver.Memory)
-	if !ok {
-		// For an unexpected reason we are not dealing with the memory storage driver.
-		return
-	}
-
-	actionConfig.KubeClient = &kubefake.PrintingKubeClient{Out: io.Discard}
-
-	for _, path := range filePaths {
-		b, err := os.ReadFile(path)
-		if err != nil {
-			log.Fatal("Unable to read memory driver data", err)
-		}
-
-		releases := []*release.Release{}
-		if err := yaml.Unmarshal(b, &releases); err != nil {
-			log.Fatal("Unable to unmarshal memory driver data: ", err)
-		}
-
-		for _, rel := range releases {
-			if err := store.Create(rel); err != nil {
-				log.Fatal(err)
-			}
-		}
-	}
-	// Must reset namespace to the proper one
-	mem.SetNamespace(settings.Namespace())
 }
